requirements.txt
feedparser==6.0.11
google-generativeai==0.8.3
pytz==2024.1
requests==2.32.3